requirements.txt
selenium===4.25.0
pytest==8.3.3
ruff==0.6.9
allure-pytest==2.13.5
allure-python-commons==2.13.5
pytest-xdist==3.6.1
Faker==36.1.1
requests==2.32.3
pydantic==2.10.2
pydantic_core==2.27.1
annotated-types==0.7.0